chore: rename UniqueIdMixin to UniqueIdMixinFactory
