+++ components/lib/fieldset/Fieldset.vue
@@ -40,12 +40,12 @@ import MinusIcon from 'primevue/icons/minus';
 import PlusIcon from 'primevue/icons/plus';
 import Ripple from 'primevue/ripple';
 import BaseFieldset from './BaseFieldset.vue';
-import { UniqueIdMixin } from 'primevue/utils';
+import { UniqueIdMixinFactory } from 'primevue/utils';
 
 export default {
     name: 'Fieldset',
     extends: BaseFieldset,
-    mixins: [UniqueIdMixin()],
+    mixins: [UniqueIdMixinFactory()],
     inheritAttrs: false,
     emits: ['update:collapsed', 'toggle'],
     data() {
