+++ components/lib/utils/Utils.js
@@ -4,7 +4,7 @@ import EventBus from './EventBus';
 import HelperSet from './HelperSet';
 import ObjectUtils from './ObjectUtils';
 import UniqueComponentId from './UniqueComponentId';
-import UniqueIdMixin from './UniqueIdMixin';
+import UniqueIdMixinFactory from './UniqueIdMixinFactory';
 import ZIndexUtils from './ZIndexUtils';
 
-export { ConnectedOverlayScrollHandler, DomHandler, EventBus, HelperSet, ObjectUtils, UniqueComponentId, ZIndexUtils, UniqueIdMixin };
+export { ConnectedOverlayScrollHandler, DomHandler, EventBus, HelperSet, ObjectUtils, UniqueComponentId, ZIndexUtils, UniqueIdMixinFactory };
